+++ tests/TestCase/Driver/DriverTest.php
@@ -18,8 +18,6 @@ namespace DatabaseBackup\Test\TestCase\Driver;
 
 use DatabaseBackup\Driver\Driver;
 use DatabaseBackup\TestSuite\TestCase;
-use Symfony\Component\Process\Exception\ProcessTimedOutException;
-use Symfony\Component\Process\Process;
 
 /**
  * DriverTest class.
@@ -67,65 +65,5 @@ class DriverTest extends TestCase
         $this->assertIsArrayNotEmpty($this->Driver->getConfig());
         $this->assertNotEmpty($this->Driver->getConfig('name'));
         $this->assertNull($this->Driver->getConfig('noExistingKey'));
-    }
-
-    /**
-     * @test
-     * @uses \DatabaseBackup\Driver\Driver::export()
-     */
-    public function testExportOnFailure(): void
-    {
-        $expectedError = 'mysqldump: Got error: 1044: "Access denied for user \'root\'@\'localhost\' to database \'noExisting\'" when selecting the database';
-
-        $this->expectExceptionMessage('Export failed with error message: `' . $expectedError . '`');
-        $Driver = $this->getMockForAbstractDriverWithErrorProcess($expectedError);
-        $Driver->export($this->getAbsolutePath('example.sql'));
-    }
-
-    /**
-     * Test for `export()` method, exceeding the timeout
-     * @see https://symfony.com/doc/current/components/process.html#process-timeout
-     * @test
-     * @uses \DatabaseBackup\Driver\Driver::_exec()
-     * @uses \DatabaseBackup\Driver\Driver::export()
-     */
-    public function testExportExceedingTimeout(): void
-    {
-        $this->expectException(ProcessTimedOutException::class);
-        $this->expectExceptionMessage('The process "dir" exceeded the timeout of 60 seconds');
-        $ProcessTimedOutException = new ProcessTimedOutException(Process::fromShellCommandline('dir'), 1);
-        $Driver = $this->getMockForAbstractDriver(['_exec']);
-        $Driver->method('_exec')->willThrowException($ProcessTimedOutException);
-        $Driver->export($this->getAbsolutePath('example.sql'));
-    }
-
-    /**
-     * Test for `import()` method on failure
-     * @test
-     * @uses \DatabaseBackup\Driver\Driver::import()
-     */
-    public function testImportOnFailure(): void
-    {
-        $expectedError = 'ERROR 1044 (42000): Access denied for user \'root\'@\'localhost\' to database \'noExisting\'';
-        $this->expectExceptionMessage('Import failed with error message: `' . $expectedError . '`');
-        $Driver = $this->getMockForAbstractDriverWithErrorProcess($expectedError);
-        $Driver->import($this->getAbsolutePath('example.sql'));
-    }
-
-    /**
-     * Test for `import()` method, exceeding the timeout
-     * @see https://symfony.com/doc/current/components/process.html#process-timeout
-     * @test
-     * @uses \DatabaseBackup\Driver\Driver::_exec()
-     * @uses \DatabaseBackup\Driver\Driver::import()
-     */
-    public function testImportExceedingTimeout(): void
-    {
-        $this->expectException(ProcessTimedOutException::class);
-        $this->expectExceptionMessage('The process "dir" exceeded the timeout of 60 seconds');
-        $ProcessTimedOutException = new ProcessTimedOutException(Process::fromShellCommandline('dir'), 1);
-        $Driver = $this->getMockForAbstractDriver(['_exec']);
-        $Driver->method('_exec')->willThrowException($ProcessTimedOutException);
-        $Driver->import($this->getAbsolutePath('example.sql'));
     }
 }